backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
# psycopg2-binary==2.9.9  # Commented out - using SQLite for development

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0

# Additional utilities
email-validator==2.1.0

# Payment Processing
stripe==7.6.0

# Email & SMS
twilio==8.10.3
jinja2==3.1.2